pyproject.toml
[tool.poetry]
name = "sqlalchemy-tree-demo"
version = "0.1.0"
description = ""
authors = ["nikita <ndavydovskij@webiomed.ai>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.9"
asyncpg = "^0.22.0"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-asyncio = "^0.15"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
